move fuse / mount cmds tests to own module
